refactor(front): Rename admin menu component
